requirements-full.txt
# requirements-full.txt
# Полный список всех зависимостей для Genesis Trading System

# Основа и GUI
PySide6>=6.8.0
pyqtgraph>=0.13.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=42.0.0

# Данные и анализ
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
scipy>=1.11.0

# Торговля
MetaTrader5>=5.0.0
arch>=6.0.0

# ML/AI
torch>=2.5.1
torchvision>=0.20.0
torchaudio>=2.5.0
lightgbm>=4.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
optuna>=3.0.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
shap>=0.42.0
numba>=0.58.0

# Базы данных
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Векторные БД и кэш
qdrant-client>=1.3.0
redis>=4.5.0
aioredlock>=0.7.0

# Web
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=10.0
aiohttp>=3.9.0
httpx>=0.24.0

# Парсинг и данные
feedparser>=6.0.0
ntscraper>=0.3.0
telethon>=1.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Утилиты
requests>=2.32.0
urllib3>=1.26.0
pysocks>=1.7.1
keyring>=23.0.0
schedule>=1.2.0
slowapi>=0.1.9
dependency-injector>=4.40.0
jinja2>=3.1.0
click>=8.0.0

# Тестирование
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-qt>=4.2.0
